Move relational.rs
